backend/requirement.txt
django-cors-headers
Pillow
djangorestframework
djangorestframework-simplejwt
mysqlclient
